move some functions to requests.go
